refactor: 歩数トップ10のリストのウィジェットを別ファイルにした & フォルダ構成の若干の変更
